% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_uu.R
\name{theme_uu}
\alias{theme_uu}
\title{ggplot2 Ulster University PhD theme}
\usage{
theme_uu(...)
}
\arguments{
\item{...}{Additional arguments to pass to ggplot2::theme()}
}
\value{
plot matching PhD styling
}
\description{
A ggplot2 theme for my PhD thesis based on ggpubr::pubr() function
}
\examples{
ggplot2::ggplot(data = cars, ggplot2::aes(x = speed, y = dist)) +
ggplot2::geom_point() +
theme_uu()

}
