% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_ulster.R
\name{ulster_cols}
\alias{ulster_cols}
\title{Function to extract ulster colors as hex codes}
\usage{
ulster_cols(...)
}
\arguments{
\item{...}{Character names of ulster_colour}
}
\description{
Function to extract ulster colors as hex codes
}
