% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_color_ulster.R
\name{ulster_pal}
\alias{ulster_pal}
\title{Return function to interpolate a ulster color palette}
\usage{
ulster_pal(palette = "main_1", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in ulster_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
Return function to interpolate a ulster color palette
}
